% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavioural_traits.R
\name{calc_area}
\alias{calc_area}
\title{calc_area()}
\usage{
calc_area(xy, arena.df)
}
\arguments{
\item{xy}{Transformed Zantiks coordinates.}

\item{arena.df}{A small dataframe containing the coordinates of the arenas within the Zantiks tank/enclosure.
This should include xmin, xmax, ymin, ymax for each arena.}

\item{ID}{a logical vector. If true, an ID will be assigned from the Service part of the file.}
}
\value{
A dataframe containing arena measurements for each individual.
}
\description{
This function estimates the proportion of the arena that covered
using the xy coordinates provided by Zantiks.
}
