% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_plots.R
\name{density_plot}
\alias{density_plot}
\title{density_plot()}
\usage{
density_plot(XY, track)
}
\arguments{
\item{XY}{coord data from a single individual}

\item{track}{Logical vector to show the track on the plot or not}
}
\value{
A plot showing the regions with the highest density
}
\description{
This function uses xy data to create a plot showing how used each area of
the area was used by an individual.
}
