% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_area.R
\name{estimate_RA}
\alias{estimate_RA}
\title{estimate_RA()}
\usage{
estimate_RA(dir, data, arena.df)
}
\arguments{
\item{dir}{A directory containing the xy data you wish to sample form}

\item{data}{A dataframe containing behavioral data from summary_behaviour()}

\item{arena.df}{A small dataframe containing the coordinates of the arenas within the Zantiks tank/enclosure.
This should include xmin, xmax, ymin, ymax for each arena.}
}
\value{
The original dataframe with the addition of relative area.
}
\description{
This function uses a random sample of xy OFT data to generate random swims to estimate
the expected area covered for a given tracklength. If you use this variable please cite:
Houslay et al. 2022, https://doi.org/10.7554/eLife.67126
}
