% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavioural_traits.R
\name{freezings}
\alias{freezings}
\title{freezings()}
\usage{
freezings(data, ID, frz)
}
\arguments{
\item{data}{Zantiks transformed csv}

\item{ID}{(optional) A logical vector. If true, the ID will be carried forward.}

\item{frz}{The number of seconds after which an animal can be considered to give the freezing response.}
}
\value{
Summary of the freezing including total number of freezing and average number of freezing.
}
\description{
This function estimates whether a stopping is actually a freezing event and
calculates the number of freezing events and the average time spent frozen.
}
