% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reallocate_zones.R
\name{is_point_in_zone}
\alias{is_point_in_zone}
\title{is_point_in_zone}
\usage{
is_point_in_zone(x, y, xmin, xmax, ymin, ymax)
}
\arguments{
\item{x}{X coordinate of to be checked}

\item{y}{Y coordinate to be checked}

\item{xmin}{Lower X boundary of the zone}

\item{xmax}{Upper X boundary of the zone}

\item{ymin}{Lower Y boundary of the zone}

\item{ymax}{Upper Y boundry of the Zone}
}
\description{
This function is used to determine whether a point is within a zone or not.
}
