% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_zantiks.R
\name{read_manyzancoord}
\alias{read_manyzancoord}
\title{read_manyzancoord()}
\usage{
read_manyzancoord(dir)
}
\arguments{
\item{dir}{Directory containing the coordinate files to be imported}
}
\value{
A dataframe containing the coord data
}
\description{
This function loads in a many .coord output files from Zanticks.
}
