% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_zanticks.R
\name{read_manyzancsv}
\alias{read_manyzancsv}
\title{read_
This function loads in all Zanticks output .csv files stored in a single directory.
It identifies the header line and only reads in from there
skipping the preamble.}
\usage{
read_manyzancsv(dir)
}
\arguments{
\item{dir}{Path to the directory where csv to be imported are stored}
}
\value{
A dataframe containing the csv data
}
\description{
read_
This function loads in all Zanticks output .csv files stored in a single directory.
It identifies the header line and only reads in from there
skipping the preamble.
}
