% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_zanticks.R
\name{read_zancoord}
\alias{read_zancoord}
\title{read_zancoord()}
\usage{
read_zancoord(dir, file)
}
\arguments{
\item{dir}{Directory containing the coordinate file to be imported}

\item{file}{Name of file to the coordinate file to be imported}
}
\value{
A dataframe containing the coord data
}
\description{
This function loads in a single .coord file output from Zanticks.
}
