% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_zantiks.R
\name{read_zancsv}
\alias{read_zancsv}
\title{read_zancsv()}
\usage{
read_zancsv(file, ID)
}
\arguments{
\item{file}{Path to the file to be imported}

\item{ID}{(optional) The location to look at to find identifying information.
Either "Service" or "Subject".}
}
\value{
A dataframe containing the csv data
}
\description{
This function loads in a single .csv file output from Zantiks.
It identifies the header line and only reads in from there
skipping the preamble.
}
