% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_plots.R
\name{track_plot}
\alias{track_plot}
\title{track_plot()}
\usage{
track_plot(XY, zones)
}
\arguments{
\item{XY}{coord data from a single individual}

\item{zones}{A dataframe with the number of zones and the coordinates of the zones.
e.g. zone, xmin, xmax, ymin, ymax, colour}
}
\value{
An object containing a longer transposed version of the data frame
}
\description{
This function uses xy data to create a plot of the track traveled by
the individual.
}
