% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data.R
\name{transform_csv}
\alias{transform_csv}
\title{transform_csv()}
\usage{
transform_csv(data, ID)
}
\arguments{
\item{data}{Dataframe to transform}

\item{ID}{a logical vector. If true, ID will be retained in the output}
}
\value{
An object containing a longer transposed version of the data frame
}
\description{
This function takes the wide dataframe and makes it longer so each row
then reflects what is occuring in one arena at that time point.
}
