% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data.R
\name{transform_xy}
\alias{transform_xy}
\title{transform_xy()}
\usage{
transform_xy(data)
}
\arguments{
\item{data}{Dataframe to transform}
}
\value{
An object containing a longer transposed version of the data frame
}
\description{
This function takes the wide xy coordinate file and makes it longer so
each row reflects what is occurring in one arena.
}
