% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_area.R
\name{walk_dist}
\alias{walk_dist}
\title{walk_dist()}
\usage{
walk_dist(
  tracklength = 1132,
  xlim = c(0, 360),
  ylim = c(0, 270),
  start = c(180, 67.5),
  stepsize = df_step_dist_cm
)
}
\arguments{
\item{xlim}{(optional) The width of the tank used for the OFT}

\item{ylim}{(optional) The length of the tank used for the OFT}

\item{start}{(optional) Starting position for the swim in x,y format}

\item{stepsize}{A dataframe containing steps estimated from real data.}

\item{data}{A dataframe containing behavioural data from summary_behaviour()}
}
\value{
An data frame containing x,y coordinates for a random swim.
}
\description{
This function estimates random swims (or walks if you have legs). This can then be used
to generate relative area. It works within the estimate_RA() fucntion
}
