% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_zanticks.R
\name{read_zancsv}
\alias{read_zancsv}
\title{read_zancsv()}
\usage{
read_zancsv(file, ID)
}
\arguments{
\item{file}{Path to the file to be imported}

\item{ID}{a locgical vector. If true, an ID will be assigned from the Service part of the file.}
}
\value{
A dataframe containing the csv data
}
\description{
This function loads in a single .csv file output from Zanticks.
It identifies the header line and only reads in from there
skipping the preamble.
}
