% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavioural_traits.R
\name{summary_behaviour}
\alias{summary_behaviour}
\title{summary_behaviour()}
\usage{
summary_behaviour(data, xy, arena.df, ID, frz)
}
\arguments{
\item{data}{Zantiks transformed csv}

\item{xy}{Zantiks transformed xy coords}

\item{arena.df}{A small dataframe containing the coordinates of the arenas within the Zantiks tank/enclosure.
This should include xmin, xmax, ymin, ymax for each arena.}

\item{ID}{a logical vector. If true, an ID will be assigned from the Service part of the file.}

\item{frz}{A threshold value after which individuals should be considered to be exhibiting a freeze response.}
}
\value{
A dataframe containing summary behavioural variables
}
\description{
This function estimates the summary behaviour data from the
zantiks csv. The variables included are time in each zone,
overall velocity, track length, freezings (calculated as no
movement for three seconds)
}
