% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_plots.R
\name{track_plot}
\alias{track_plot}
\title{track_plot()}
\usage{
track_plot(XY, zones, nested)
}
\arguments{
\item{XY}{coord data from a single individual}

\item{zones}{A dataframe with the number of zones and the coordinates of the zones.
e.g. zone, xmin, xmax, ymin, ymax, group
Group could be light/dark or inner and outer in nested}

\item{nested}{Logical vector is Zone1 within Zone2?}
}
\value{
A plot showing the complete path travelled by individuals
}
\description{
This function uses xy data to create a plot of the track traveled by
the individual.
}
